+++ demo/src/main/webapp/js/blutorange.js
@@ -1,4 +1,3 @@
-
 class Blutorange {
 	constructor()  {
 		this.version = "final";