+++ src/main/java/com/github/blutorange/maven/plugin/closurecompiler/common/AggregationType.java
@@ -1,8 +1,0 @@
-package com.github.blutorange.maven.plugin.closurecompiler.common;
-
-/**
- * Defines the aggregation type.
- */
-public enum AggregationType {
-js
-}